Add coverage report artifact to presubmit workflow
It's frequently helpful to be able to see the coverage report for a
given commit. To enable this, the previous `presubmit` workflow has been
modified to now include a step that generates a coverage report and
uploads it as an artifact.

Artifacts are taken from all 3 OS runs, and are added as an artifact
output. Intermediate artifacts expire in 1 day to reduce retention,
but the final coverage report is retained for 7 days.

In the process, the `presubmit` workflow was modified to be able to be
called from a separate pipeline so that this may be connected into
the soon-to-be-added deployment pipeline, as well as the docs
deployment pipeline.

Change-Category: infrastructure