Add basic Go devcontainer
Devcontainers are useful for encapsulating the set of dependencies and
tools required to get going and do basic work. Thankfully, there are
plenty of pre-bundled ones by Microsoft we can use for this purpose!

This adds in a `devcontainer` for doing Go development that can be used
for this project.

Change-Category: infrastructure