Add dedent package for strings
Go lacks a package for dedenting strings, which makes it a problem when
you want to write long multiline strings in a way that won't also be
output verbatim from tooling. As such, this is a necessity for working
with tools.

This adds a basic dedent functionality as its own package, with some
tests.

Change-Category: library