Refactor GitHub Actions workflows
The GitHub Actions workflows created earlier today were put together in
a single workflow file for ease initially; but it's getting to a point
where it would be easier to have access to reusable workflows.

This breaks apart the workflows so that linting and building can be used
independently for deployment purposes. In the process, the `presubmit`
was renamed `continuous-integration` since `presubmit` would otherwise
run technically after post-submit.

Change-Category: integration