Add Mermaid plugin to devcontainer
The devcontainer was missing the mermaid plugin for markdown rendering,
so this commit adds it in. In the process, the JSON is now being
reformatted to use tabs for some reason, so more lines are showing as
changed than actually are.

This also adds a few more basic settings to the default settings.

Change-Category: infrastructure