Fix stalling GitHub workflow
Take-two on fixing this issue.
I installed `godoc`, but didn't realize `setup-go` was never called,
so this means that `GOPATH` is never set up and the `bin` directory
isn't in `$PATH`. The consequence is that `godoc` can't run.
This attempts again to fix it.

Additionally, the script has been updated to `euo pipefail` to ensure
it won't happen again.

Change-Category: integration
Change-Type: fix