Add `dedent.Strings` for dedenting multiple lines
Previously this only included `dedent.String` for dedenting a single
string that is treated as multiple lines and re-joined. However it's
also frequently desirable to pass N strings that will be each treated
as lines, and will be dedented accordingly.

This commit adds `dedent.Strings` for this purpose.

Change-Category: library