Fix CodeQL warning for Go installation
CodeQL is generating warnings that Go is being installed after CodeQL is
setup, which apparently may result in incorrect behavior for some
reason. This warning is surprisingly meta, since the warning is about
the system that should be generating warnings.

This fixes the underlying issue by installing Go before the CodeQL check
is set up.

Change-Category: infrastructure
Change-Type: fix