Remove actions/cache from github workflows
For some reason, the actions-cache is failing to be retrieved or
updated due to permission issues. Evidently this still builds fast
enough without it being present, and so this is clearly not a
botteneck for this project (yet); so it's just being removed for now
rather than spending cycles to analyze why. This can be revisited in
the future.

Change-Category: infrastructure
Change-Type: fix