Add package for globbing support
The builtin go functionality for glob support is quite minimal,
and only supports trivial shell-substitutions and wildcard syntax --
but does not support multi-path matching, or negation.

This logic will be necessary for finding all `*.proto` source files
that will be processed by the `protoc` compiler in an easy way, so
this adds a new package that does this.

Change-Category: library