Move flagset into `cli` subdirectory
This is just a minor restructuring of an internal package;
I'm realizing now that `flagset` is logically only associated to anything
using a `cobra.Command`, so it would be way easier to keep all things
together. From now on, the `cli` package will bucket utilities that
are in terms of spf13's `pflag` or `flagset` logic so that it's more
focused.

Other subpackages will be focused around their unique functionalities.